Make the swisscard importer work with the new statement CSV format.
Somewhen between Jun 4th and Jul 5th, the swisscard statements in CSV
format moved from a single Category column to:

- a "Merchant Category" column, which looks like a general category for
the merchant.
- a "Registered Category" column, which looks like a string version
of the ISO standard MCC code.

Since all statements, previous and past, are now issued with the updated
format, it's safe to just modify the importer to make use of the split
columns as metadata.